+++ lib/core/src/display/mesh_registry.rs
@@ -3,8 +3,8 @@ use crate::prelude::*;
 use crate::dirty;
 use crate::dirty::traits::*;
 use crate::data::function::callback::*;
-use crate::display::symbol::attribute as attr;
-use crate::display::symbol::attribute::item::Item;
+use crate::display::symbol::buffer as attr;
+use crate::display::symbol::buffer::item::Item;
 use crate::display::symbol::mesh;
 use crate::system::web::fmt;
 use crate::system::web::group;
