+++ lib/core/src/display/symbol/mod.rs
@@ -1,4 +1,4 @@
-pub mod attribute;
+pub mod buffer;
 pub mod geometry;
 pub mod mesh;
 pub mod scope;
